+++ src/api/routes.ts
@@ -156,6 +156,12 @@ export const setupRoutes = (server: Server) => {
     options: adminEndpoints.postSetCollectionCommunity,
   });
 
+  server.route({
+    method: "GET",
+    path: "/admin/get-marketplaces",
+    options: adminEndpoints.getMarketplaces,
+  });
+
   // API keys
 
   server.route({
@@ -497,6 +503,12 @@ export const setupRoutes = (server: Server) => {
   });
 
   server.route({
+    method: "GET",
+    path: "/liquidity/users/v2",
+    options: ordersEndpoints.getUsersLiquidityV2Options,
+  });
+
+  server.route({
     method: "POST",
     path: "/order/v1",
     options: ordersEndpoints.postOrderV1Options,
@@ -660,6 +672,12 @@ export const setupRoutes = (server: Server) => {
     method: "GET",
     path: "/users/{user}/tokens/v2",
     options: tokensEndpoints.getUserTokensV2Options,
+  });
+
+  server.route({
+    method: "GET",
+    path: "/users/{user}/tokens/v3",
+    options: tokensEndpoints.getUserTokensV3Options,
   });
 
   server.route({