+++ src/orderbook/orders/index.ts
@@ -15,4 +15,7 @@ export type OrderKind =
   | "opendao-erc1155"
   | "foundation"
   | "x2y2"
-  | "seaport";
+  | "seaport"
+  | "rarible"
+  | "element-erc721"
+  | "element-erc1155";