+++ src/sync/events/data/index.ts
@@ -11,6 +11,8 @@ import * as wyvernV2 from "@/events-sync/data/wyvern-v2";
 import * as wyvernV23 from "@/events-sync/data/wyvern-v2.3";
 import * as x2y2 from "@/events-sync/data/x2y2";
 import * as zeroExV4 from "@/events-sync/data/zeroex-v4";
+import * as rarible from "@/events-sync/data/rarible";
+import * as element from "@/events-sync/data/element";
 
 // All events we're syncing should have an associated `EventData`
 // entry which dictates the way the event will be parsed and then
@@ -48,7 +50,12 @@ export type EventDataKind =
   | "x2y2-order-inventory"
   | "seaport-order-cancelled"
   | "seaport-order-filled"
-  | "seaport-counter-incremented";
+  | "seaport-counter-incremented"
+  | "rarible-match"
+  | "element-erc721-sell-order-filled"
+  | "element-erc721-buy-order-filled"
+  | "element-erc1155-sell-order-filled"
+  | "element-erc1155-buy-order-filled";
 
 export type EventData = {
   kind: EventDataKind;
@@ -92,6 +99,11 @@ export const getEventData = (eventDataKinds: EventDataKind[] | undefined) => {
       zeroExV4.erc1155OrderFilled,
       x2y2.orderCancelled,
       x2y2.orderInventory,
+      rarible.match,
+      element.erc721BuyOrderFilled,
+      element.erc721SellOrderFilled,
+      element.erc1155BuyOrderFilled,
+      element.erc1155SellOrderFilled,
     ];
   } else {
     return (
@@ -168,6 +180,16 @@ const internalGetEventData = (kind: EventDataKind): EventData | undefined => {
       return seaport.orderCancelled;
     case "seaport-order-filled":
       return seaport.orderFulfilled;
+    case "rarible-match":
+      return rarible.match;
+    case "element-erc721-sell-order-filled":
+      return element.erc721SellOrderFilled;
+    case "element-erc721-buy-order-filled":
+      return element.erc721BuyOrderFilled;
+    case "element-erc1155-sell-order-filled":
+      return element.erc1155SellOrderFilled;
+    case "element-erc1155-buy-order-filled":
+      return element.erc1155BuyOrderFilled;
     default:
       return undefined;
   }